// src/utils/types/user.ts
// types/user.ts

import { BaseRecord } from ".";

interface Coordinates {
    lat: number;
    lng: number;
  }
  
  interface Address {
    address: string;
    city: string;
    state: string;
    stateCode: string;
    postalCode: string;
    coordinates: Coordinates;
    country: string;
  }
  
  interface Hair {
    color: string;
    type: string;
  }
  
  interface Bank {
    cardExpire: string;
    cardNumber: string;
    cardType: string;
    currency: string;
    iban: string;
  }
  
  interface Company {
    department: string;
    name: string;
    title: string;
    address: Address;
  }
  
  interface Crypto {
    coin: string;
    wallet: string;
    network: string;
  }
  
  export interface User  extends BaseRecord {
    firstName: string;
    lastName: string;
    maidenName: string;
    age: number;
    gender: 'male' | 'female' | 'other';
    email: string;
    phone: string;
    username: string;
    password: string;
    birthDate: string;
    image: string;
    bloodGroup: string;
    height: number;
    weight: number;
    eyeColor: string;
    hair: Hair;
    ip: string;
    address: Address;
    macAddress: string;
    university: string;
    bank: Bank;
    company: Company;
    ein: string;
    ssn: string;
    userAgent: string;
    crypto: Crypto;
    role: 'admin' | 'moderator' | 'user';
    status: 'active' | 'inactive';
  }
  
  // You can also create a type for the users array
  export interface UserResponse {
    users: User[];
    total: number;
    skip: number;
    limit: number;
}